Make a templatetag which will render & process the form.
This also requires a contextprocessor to be enabled to put the request
object into the context.

We also add state to the AbstractForm object, so that the generation of
entries & the sending of signals only happens once per process, even
if the form gets processed multiple times for any reason.