Add max length checking
As we are appending data to the end of a slug it adds additional length
to the slug. This can mean we have gone past the fields length causing
database errors.

As we are passing a manager and the field name into the function we can
determine the maximum length of the field and limit the slug to that
length.